+++ crate/cfg_model/src/rt.rs
@@ -1,18 +1,15 @@
+//! Runtime data types referenced within configuration.
+
 pub use self::{
-    check_status::CheckStatus,
-    files::{Files, RwFiles},
-    progress_limit::ProgressLimit,
-    station::Station,
-    station_errors::StationErrors,
-    station_mut::StationMut,
-    station_progress::StationProgress,
-    station_rt_id::StationRtId,
-    train_report::TrainReport,
+    check_status::CheckStatus, files::Files, files_rw::FilesRw, progress_limit::ProgressLimit,
+    station::Station, station_errors::StationErrors, station_mut::StationMut,
+    station_progress::StationProgress, station_rt_id::StationRtId, train_report::TrainReport,
     visit_status::VisitStatus,
 };
 
 mod check_status;
 mod files;
+mod files_rw;
 mod progress_limit;
 mod station;
 mod station_errors;
